requirements.txt
absl-py==0.15.0
asgiref==3.4.1
astunparse==1.6.3
bert-score==0.3.10
bpemb==0.3.3
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.8
click==8.0.3
conllu==4.4.1
cycler==0.11.0
datasets==1.2.1
Deprecated==1.2.13
dill==0.3.4
Django==3.2.9
docopt==0.6.2
editdistance==0.6.0
filelock==3.4.0
fire==0.4.0
flair==0.10
flatbuffers==1.12
fonttools==4.28.2
ftfy==6.0.3
gast==0.4.0
gdown==3.12.2
gensim==4.1.2
google-auth==2.3.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.34.1
h5py==3.1.0
huggingface-hub==0.2.0
idna==3.3
importlib-metadata==4.8.2
Janome==0.4.1
joblib==1.1.0
keras==2.7.0
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
keras2onnx==1.7.0
kiwisolver==1.3.2
konoha==4.6.3
langdetect==1.0.9
language-tool-python==2.6.2
lemminflect==0.2.2
libclang==12.0.0
lru-dict==1.1.7
lxml==4.6.4
Markdown==3.3.6
matplotlib==3.5.0
more-itertools==8.8.0
mpld3==0.3
multiprocess==0.70.12.2
nltk==3.6.5
num2words==0.5.10
numpy==1.19.5
oauthlib==3.1.1
onnx==1.10.2
onnxconverter-common==1.8.1
opt-einsum==3.3.0
overrides==3.1.0
packaging==21.3
pandas==1.3.4
Pillow==8.4.0
protobuf==3.19.1
pyarrow==6.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.6
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
regex==2021.11.10
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.8
sacremoses==0.0.46
scikit-learn==1.0.1
scipy==1.7.3
segtok==1.5.10
sentencepiece==0.1.95
setuptools-scm==6.3.2
six==1.15.0
smart-open==5.2.1
sqlitedict==1.7.0
sqlparse==0.4.2
tabulate==0.8.9
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorboardX==2.4.1
tensorflow-cpu==2.7.0
tensorflow-estimator==2.7.0
tensorflow-hub==0.12.0
tensorflow-io-gcs-filesystem==0.22.0
tensorflow-text==2.5.0
termcolor==1.1.0
terminaltables==3.1.0
textattack==0.3.4
threadpoolctl==3.0.0
tokenizers==0.10.3
tomli==1.2.2
torch==1.10.0
tqdm==4.49.0
transformers==4.12.5
typing-extensions==3.7.4.3
urllib3==1.26.7
wcwidth==0.2.5
Werkzeug==2.0.2
Wikipedia-API==0.5.4
word2number==1.1
wrapt==1.12.1
xxhash==2.0.2
zipp==3.6.0
